requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.1
djangorestframework==3.13.1
filelock==3.6.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
